services/cyberdesk-operator/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
durationpy==0.9
frozenlist==1.5.0
google-auth==2.38.0
idna==3.10
iso8601==2.1.0
kopf==1.37.5
kubernetes==32.0.1
multidict==6.4.3
oauthlib==3.2.2
propcache==0.3.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.3
pydantic-core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
pyyaml==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
websocket-client==1.8.0
yarl==1.19.0
